% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A Reference Class for cumputating Linear Regression}
\arguments{
\item{formula}{A formula}

\item{data}{A data frame}
}
\description{
This class has a vary of functions.
}
\section{Fields}{

\describe{
\item{\code{computing}}{a liner regression 
library(ggplot2)}
}}


