% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame containing the variables in the model.}
}
\value{
A list containing model coefficients and residuals.
}
\description{
This function implements a basic linear regression model.
}
\examples{
data(mtcars)
fit <- my_lm(mpg ~ wt + hp, data = mtcars)
print(fit$coefficients)
}
