% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{myLinearData}
\alias{myLinearData}
\title{A simple dataset}
\format{
A data frame with 100 rows and 7 variables:
\describe{
\item{Y}{The average climate of a place}
\item{X1}{Average height from sea level}
\item{X2}{Average distance from equator}
\item{X3}{Average distance from sea}
\item{X4}{Average rainfall}
...
}
}
\usage{
myLinearData
}
\description{
A description of a simple dataset
}
\keyword{datasets}
