% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myLinearRegression.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Performs linear regression.}
\usage{
myLinearRegression(Y, covData)
}
\arguments{
\item{Y}{vector of outcomes.}

\item{covData}{list of subjects (data corresponding to covariates)}
}
\value{
Coefficients \code{coeff} and p-values \code{pval}.
}
\description{
This is truly a great and much-needed function
}
\examples{
myLinearRegression(y,covData)
}
